Backend/requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
rest_framework_simplejwt==0.0.2